Moved Calculator.js and Flavors.js to correct location
